+++ sync-for-payroll/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.1.0] sync-for-payroll
 ### Releases
 - [NPM v1.1.0] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.1.0 - sync-for-payroll
+
+## 2023-09-21 00:05:23
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.88.1 (2.122.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.1] sync-for-payroll
+### Releases
+- [NPM v1.1.1] https://www.npmjs.com/package/@codat/sync-for-payroll/v/1.1.1 - sync-for-payroll